ripple/requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
gitpython>=3.1.40
networkx>=3.2
python-dotenv>=1.0.0

